fix bluetooth headsets on iOS
Something recently stopped working and upon investigation, it seems that
it's related to the buffer duration. Finding the Goldilocks zone for
duration was a mix of trial and error, some thinking about latency, and
the likelihood of < 44.1khz sampling rates.

At 48khz, 3072 samples takes 64ms. At At 44.1khz, 70ms. 32khz is 96ms.